feat(DTFS2-7055): handle feature flag in deleteOne helper
