+++ libs/common/src/change-stream/test-helpers/index.ts
@@ -1,2 +1,2 @@
 export * from './generate-mock-audit-database-record';
-export * from './with-deletion-audit-logs.api-tests';
+export * from './with-delete-one.api-tests';
